requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
anthropic==0.40.0
python-dotenv==1.0.1
weasyprint==68.1
xhtml2pdf==0.2.17
jinja2==3.1.4
psycopg2-binary==2.9.12
playwright==1.59.0
